Better and simpler partial fix to issue 90
If page_data.figures is empty (or not agreeing with images_per_page at all,
either above or below), it really means the HNParsePage code is unable to
extract the image offsets and sizes correctly. In this case, we should
simply output the images individually.

This is a much simpler answer to trying to truncate the list of co-ordinates
to match number of found images, as in the previous code. That also could not
cope if list of co-ordinates is smaller than found images.