Check that first image of multi-image page is indeed offset'ed from corner.
There are text-only CAJ files with multi-images, none of which is the
background. This problem is seen with issue 66,
IDL编译器的实现_词法分析部分_于埴尧.caj .

Before this change, the 2nd image on page 3 is output as
(1bit jpeg!? "Syntax Error (8584): Unknown DCT marker <31>")
[ DCT ] 666x172 1bpc DevRGB
instead of
[ Flate ] 666x172 1bpc Idx too.

There is probably a bug/assumption on secondary images being jpeg
in the multi-image page code.